refactor(app): separate Express app from server startup
- Extract app configuration to app.ts for testability
- Keep server startup logic in index.ts
- Export app instance for use in tests
