+++ api/src/app.ts
@@ -1,0 +1,31 @@
+import express, { Request, Response } from 'express';
+import helmet from 'helmet';
+import routes from './routes';
+import { corsMiddleware } from './middlewares/cors';
+import { logger } from './middlewares/logger';
+import { errorHandler, notFoundHandler } from './middlewares/errorHandler';
+
+const app = express();
+
+app.use(helmet());
+app.use(corsMiddleware);
+app.use(logger);
+app.use(express.json());
+app.use(express.urlencoded({ extended: true }));
+
+app.get('/', (req: Request, res: Response) => {
+  res.json({
+    message: 'English Playlist Gallery API',
+    version: '1.0.0',
+    status: 'healthy',
+    timestamp: new Date().toISOString(),
+  });
+});
+
+app.use('/api', routes);
+
+app.use(notFoundHandler);
+
+app.use(errorHandler);
+
+export default app;
